// frontend/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Any: { input: any; output: any; }
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: { input: any; output: any; }
  Duration: { input: any; output: any; }
  Map: { input: any; output: any; }
  Time: { input: any; output: any; }
  UUID: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type ActionCacheStatistics = Node & {
  __typename?: 'ActionCacheStatistics';
  actionSummary?: Maybe<ActionSummary>;
  cacheCheckSemaphoreWaitTimeInMs?: Maybe<Scalars['Int']['output']>;
  hits?: Maybe<Scalars['Int']['output']>;
  id: Scalars['ID']['output'];
  loadTimeInMs?: Maybe<Scalars['Int']['output']>;
  missDetails?: Maybe<Array<MissDetail>>;
  misses?: Maybe<Scalars['Int']['output']>;
  saveTimeInMs?: Maybe<Scalars['Int']['output']>;
  sizeInBytes?: Maybe<Scalars['Int']['output']>;
};
// ...
export type ActionData = Node & {
  __typename?: 'ActionData';
  actionSummary?: Maybe<ActionSummary>;
  actionsCreated?: Maybe<Scalars['Int']['output']>;
  actionsExecuted?: Maybe<Scalars['Int']['output']>;
  firstStartedMs?: Maybe<Scalars['Int']['output']>;
  id: Scalars['ID']['output'];
  lastEndedMs?: Maybe<Scalars['Int']['output']>;
  mnemonic?: Maybe<Scalars['String']['output']>;
  systemTime?: Maybe<Scalars['Int']['output']>;
  userTime?: Maybe<Scalars['Int']['output']>;
};
// ...
export type ActionSummary = Node & {
  __typename?: 'ActionSummary';
  actionCacheStatistics?: Maybe<ActionCacheStatistics>;
  actionData?: Maybe<Array<ActionData>>;
  actionsCreated?: Maybe<Scalars['Int']['output']>;
  actionsCreatedNotIncludingAspects?: Maybe<Scalars['Int']['output']>;
  actionsExecuted?: Maybe<Scalars['Int']['output']>;
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  remoteCacheHits?: Maybe<Scalars['Int']['output']>;
  runnerCount?: Maybe<Array<RunnerCount>>;
};
// ...
export type Metrics = Node & {
  __typename?: 'Metrics';
  actionSummary?: Maybe<ActionSummary>;
  artifactMetrics?: Maybe<ArtifactMetrics>;
  bazelInvocation?: Maybe<BazelInvocation>;
  buildGraphMetrics?: Maybe<BuildGraphMetrics>;
  cumulativeMetrics?: Maybe<CumulativeMetrics>;
  dynamicExecutionMetrics?: Maybe<DynamicExecutionMetrics>;
  id: Scalars['ID']['output'];
  memoryMetrics?: Maybe<MemoryMetrics>;
  networkMetrics?: Maybe<NetworkMetrics>;
  packageMetrics?: Maybe<PackageMetrics>;
  targetMetrics?: Maybe<TargetMetrics>;
  timingMetrics?: Maybe<TimingMetrics>;
  workerMetrics?: Maybe<Array<WorkerMetrics>>;
  workerPoolMetrics?: Maybe<WorkerPoolMetrics>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type RunnerCount = Node & {
  __typename?: 'RunnerCount';
  actionSummary?: Maybe<ActionSummary>;
  actionsExecuted?: Maybe<Scalars['Int']['output']>;
  execKind?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
};